requirements.txt
pytest
pytesseract
Pillow
